pyproject.toml
[tool.poetry]
name = "regelum-control"
package-mode = true
version = "0.3.1"
description = "Regelum is a flexibly configurable framework for agent-environment simulation with a menu of predictive and reinforcement learning pipelines."
authors = ["Georgiy Malaniya <pwlsd.gm@gmail.com> Anton Bolychev <bolychev.anton@gmail.com> Grigoriy Yaremenko <yaremenko8@gmail.com> Pavel Osinenko <p.osinenko@yandex.ru>"]
license = "MIT"
readme = "README.md"
packages = [
    { include = "regelum" },
]

[tool.poetry.dependencies]
python = "^3.9"
torch = ">=1.13.1"
numpy = ">=1.25.2"
scipy = ">=1.11.0"
omegaconf = "^2.3.0"
gitpython = "^3.1.42"
mlflow = "=2.4.1"
pandas = ">=1.5.3"
tables = ">=3.8.0"
pyqt5 = "5.15.10"
pyqt5-qt5 = ">=5.15.2,<=5.15.13"
casadi = "^3.6.5"
cues = "^0.3.0"
svgpathtools = "^1.6.1"
dill = "^0.3.8"
matplotlib = ">=3.7.1"
svgpath2mpl = "^1.0.0"
plotly = ">=5.15.0"
rehydra = "=1.3.5"
rehydra-joblib-launcher = "=1.3.0"
matplotx = "^0.3.8"

[tool.setuptools.package-data]
"*" = ["*.*"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
